requirements.txt
biopython==1.86
blinker==1.9.0
certifi==2026.1.4
click==8.3.1
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.4.1
python-dotenv==1.2.1
Werkzeug==3.1.5
